requirements.txt
python-dotenv
matplotlib
pandas
metaapi-cloud-sdk
nest_asyncio